pyproject.toml
[tool.poetry]
name = "nzshm-model-graphql-api"
version = "0.1.3"
description = ""
authors = ["NSHM <nshm@gns.cri.nz>", "Chris Chamberlain <chrisbc@artisan.co.nz>"]
license = "AGPL3"
readme = "README.md"
packages = [{include = "nzshm_model_graphql_api"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.4"
graphene-django = "^3.1.5"
django-filter = "^23.2"
python-dateutil = "^2.8.2"
werkzeug = "^2.3.7"
whitenoise = "^6.5.0"
django-extensions = "^3.2.3"
elasticsearch-dsl = ">=6.0.0,<7.0.0"
graphene-elastic = {path = "../graphene-elastic"}
django-elasticsearch-dsl = ">=6.0.0,<7.0.0"

[tool.poetry.group.dev.dependencies]
pyparsing = "^3.1.1"
pydot = "^1.4.2"

tox = "^4.2.8"
black  = { version = "^22.3"}
isort  = { version = "^5.8.0"}
flake8  = { version = "^3.9.2"}
flake8-docstrings = { version = "^1.6.0", optional = true }
pytest = "^7.4.0"
pytest-cov  = { version = "^2.12.0"}
virtualenv  = { version = "^20.2.2", optional = true}
pip  = { version = "^23.2.1", optional = true}

pre-commit = {version = "^2.12.0", optional = true}
toml = {version = "^0.10.2", optional = true}
bump2version = {version = "^1.0.1"}
mypy = "^1.5.1"
types-setuptools = "^68.1.0.1"
pytest-django = "^4.5.2"

[tool.poetry.extras]
test = [
    "pytest",
    "black",
    "isort",
    "mypy",
    "flake8",
    "flake8-docstrings",
    "pytest-cov"
    ]

dev = ["tox", "pre-commit", "virtualenv", "pip", "twine", "toml", "bump2version"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
